Refactored hubhop filter dialog so that we can reuse it for inputs too.
